refactor : 기존 Register(View)Controller 복구
