+++ app/src/main/java/com/techcourse/controller/RegisterViewController.java
@@ -1,0 +1,14 @@
+package com.techcourse.controller;
+
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import webmvc.org.springframework.web.servlet.mvc.asis.Controller;
+
+public class RegisterViewController implements Controller {
+
+    @Override
+    public String execute(final HttpServletRequest req, final HttpServletResponse res) throws Exception {
+        return "/register.jsp";
+    }
+
+}
